fixup return type of sizeInBytes:
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/sqlclient/trunk@37042 72102866-910b-0410-8b05-ffd578937521

+++ SQLClient.h
@@ -300,7 +300,7 @@ extern NSString * const SQLClientDidDisconnectNotification;
  * The exclude set is used to specify objects to exclude from the
  * calculation (to prevent recursion etc).
  */
-- (unsigned) sizeInBytes: (NSMutableSet*)exclude;
+- (NSUInteger) sizeInBytes: (NSMutableSet*)exclude;
 @end
 
 extern NSString	*SQLException;
